restructured mobile app code and fixed filename error
