+++ mobileapp/app/src/main/res/layout/activity_registration_result.xml
@@ -8,7 +8,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".RegistrationActivity">
+    tools:context=".logic.RegistrationActivity">
 
 
     <ImageView
